mydream/requirements.txt
numpy
lightgbm
uwsgi
flask
python-magic
click